backend/requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.16
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-filter==22.1
django-js-asset==1.2.2
django-mdeditor==0.1.15
django-uuslug==1.1.8
djangorestframework==3.14.0
drf-extensions==0.7.1
easy-thumbnails==2.7
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
Jinja2==2.11.1
MarkupSafe==1.1.1
pymysql==1.0.2
pycparser==2.19
python-slugify==6.1.2
requests==2.23.0
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.8
websockets==8.1
py3Fdfs==2.2.0
drf-yasg==1.21.4
nacos-sdk-python
